requirements.txt
aiogram==3.21.0
aiohttp>=3.9.0,<3.13.0
aiofiles>=23.2.1,<24.2
certifi
magic-filter
pydantic
typing-extensions
python-dotenv
telethon